+++ src/BecklynSchemaOrgBundle.php
@@ -6,6 +6,7 @@ use Becklyn\RadBundles\Bundle\BundleExtension;
 use Becklyn\SchemaOrg\Schema\MetaDataNormalizer;
 use Becklyn\SchemaOrg\SchemaBuilder\SchemaBuilderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class BecklynSchemaOrgBundle extends Bundle
@@ -13,7 +14,7 @@ class BecklynSchemaOrgBundle extends Bundle
     /**
      * @inheritDoc
      */
-    public function getContainerExtension ()
+    public function getContainerExtension () : ?ExtensionInterface
     {
         return new BundleExtension($this);
     }