+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+1.7.0
+=====
+
+*   (feature) Add support for PHP 8.
+*   (internal) Fix Symfony deprecations.
+*   (internal) Replace CircleCI with GitHub Actions.
+*   (improvement) Add missing property types and return types.
+*   (internal) Remove support for Symfony 4.4.
+*   (improvement) Add support for Symfony 6.
+
+
 1.6.0
 =====